requirements.txt
typing==3.7.4.3
munch==2.5.0
numpy==1.19.4
pandas==1.1.4
tqdm==4.54.0
torch==1.7.1
requests==2.25.0
scipy==1.5.4
Pillow==8.0.1
torchvision==0.8.2
scikit_learn==0.23.2
seaborn==0.11.0
python-dotenv==0.15.0
imageio==2.9.0
hydra-core==1.0.4
tsnecuda==0.1.1
wandb==0.10.12